refactored tablebody into separate component
